# docs/swagger.yaml
basePath: /
host: https://Backend-DianasDog.app.secoder.net
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: 特型卡片搜索系统
  version: "1.0"
paths:
  /alluser:
    get:
      description: 取得所有用户名
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": ["name1", "name2"]}'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 取得所有用户名
      tags:
      - User
  /category:
    get:
      description: 后端返回现有特型卡类型之接口
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": data}'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 取得现有特型卡类型
      tags:
      - Setting
  /data:
    get:
      description: 后端返回数据之接口
      parameters:
      - description: '特型卡名称 (如: car, poem 等)'
        in: query
        name: resource
        required: true
        type: string
      - description: 文件名
        in: query
        name: filename
        required: true
        type: string
      produces:
      - multipart/form-data
      responses:
        "200":
          description: XML 文件
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 取得数据
      tags:
      - Data
    post:
      consumes:
      - multipart/form-data
      description: 后端接收 XML 数据之接口。若为插入，则需要有 data 或 file 中的一个；若为删除，则需要 key 或 filename
        中的一个。（若同时出现则两者都删除）
      parameters:
      - description: '特型卡名称 (如: car, poem 等)'
        in: formData
        name: resource
        required: true
        type: string
      - description: 操作类型 (insert, delete)
        in: formData
        name: type
        required: true
        type: string
      - description: 数据内容
        in: formData
        name: data
        type: string
      - description: 文件名
        in: formData
        name: filename
        type: string
      - description: 文件形式上传之数据
        in: formData
        name: file
        type: file
      - description: '要删除的 key (即不含 resource 的 docid, 如: mytest@test 中的 test)'
        in: formData
        name: key
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 发送 XML 数据
      tags:
      - Data
  /dataname:
    get:
      description: 后端返回数据文件名之接口
      parameters:
      - description: '特型卡名称 (如: car, poem 等)'
        in: query
        name: resource
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 文件名
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 取得数据文件名
      tags:
      - Data
  /item:
    get:
      description: 后端返回某一条数据之接口
      parameters:
      - description: '特型卡名称 (如: car, poem 等)'
        in: query
        name: resource
        required: true
        type: string
      - description: 索引值 (默认为 item.key)
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": data}'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 取得单一数据
      tags:
      - Data
  /key:
    get:
      description: 后端返回某一数据类型键值之接口
      parameters:
      - description: '特型卡名称 (如: car, poem 等)'
        in: query
        name: resource
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": data}'
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 取得数据键值
      tags:
      - Data
  /login:
    post:
      consumes:
      - application/json
      description: 后端密码接口
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"password": "pwd"}'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 密码较验接口
      tags:
      - User
  /pattern:
    get:
      description: 后端返回配置文件之接口
      parameters:
      - description: '特型卡名称 (如: car, poem 等)'
        in: query
        name: resource
        required: true
        type: string
      - description: 下载信息类型 (intent, garbage, pattern)
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": data}'
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 取得配置文件
      tags:
      - Setting
    post:
      consumes:
      - multipart/form-data
      description: 后端接收配置文件之接口
      parameters:
      - description: 上传信息类型 (intent, garbage, pattern)
        in: query
        name: type
        required: true
        type: string
      - description: '特型卡名称 (如: car, poem 等)'
        in: query
        name: resource
        required: true
        type: string
      - description: 数据内容
        in: query
        items:
          type: string
        name: data
        required: true
        type: array
      - description: 操作类型 (insert, delete)
        in: query
        name: operation
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "successful!"}'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 发送配置文件
      tags:
      - Setting
  /search:
    post:
      consumes:
      - application/json
      description: 后端搜索系统接口
      parameters:
      - description: 要搜索的句子
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 搜索系统接口
      tags:
      - Search
  /setting:
    get:
      description: 后端返回写入行为描述之接口
      parameters:
      - description: '特型卡名称 (如: car, poem 等)'
        in: query
        name: resource
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": data}'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 取得写入行为描述
      tags:
      - Setting
    post:
      consumes:
      - multipart/form-data
      description: 后端接收写入行为描述之接口
      parameters:
      - description: '特型卡名称 (如: car, poem 等)'
        in: formData
        name: resource
        required: true
        type: string
      - description: 文件形式上传之数据
        in: formData
        name: file
        type: file
      - description: 数据内容 (JSON 格式)
        in: formData
        name: data
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 发送写入行为描述
      tags:
      - Setting
  /testes:
    post:
      consumes:
      - application/json
      description: 检查 es 数据之测试接口
      parameters:
      - description: 要搜索的句子
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: es 测试接口
      tags:
      - Search
  /useradd:
    post:
      consumes:
      - application/json
      description: 增加用户
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: userpassword
        required: true
        type: string
      - description: 用户等级
        in: query
        name: userlevel
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 增加用户接口
      tags:
      - User
  /userdelete:
    post:
      consumes:
      - application/json
      description: 删除用户
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 删除用户接口
      tags:
      - User
swagger: "2.0"